........
  r2976 | KanaSugimoto | 2014-07-15 19:06:47 +0900 (Tue, 15 Jul 2014) | 14 lines
  
  
         New Development: No
              JIRA Issue: Yes (CAS-5139)
          Ready for Test: Yes
       Interface Changes: No
  What Interface Changed: 
           Test Programs: 
    Put in Release Notes: No
               Module(s): asap.sbseparator
             Description: Fixed a bug for ifno=-1.
  
  	
........


git-svn-id: https://svn.cv.nrao.edu/svn/casa/branches/release-4_2_2@30471 ac9c0da5-131f-0410-9009-a9c195f183e6

+++ asap/src/STSideBandSep.cpp
@@ -123,7 +123,7 @@ void STSideBandSep::initshift()
   tableList_.resize(0);
 };
 
-void STSideBandSep::setFrequency(const unsigned int ifno,
+void STSideBandSep::setFrequency(const int ifno,
 				 const string freqtol,
 				 const string frame)
 {
@@ -132,7 +132,7 @@ void STSideBandSep::setFrequency(const unsigned int ifno,
   initshift();
 
   // IFNO
-  sigIfno_ = (int) ifno;
+  sigIfno_ = ifno;
 
   // Frequency tolerance
   Quantum<Double> qftol;
